GoodInternetOrComplain/requirements.txt
pip==21.3.1
python-dotenv==0.19.2
setuptools==60.1.1
speedtest-cli==2.1.3
twitter==1.19.3
